[refactor/#29] 코드리뷰 반영
