+++ app/src/main/java/sopt/org/starbucks/ui/mymenu/EditMenuUiState.kt
@@ -1,10 +1,11 @@
 package sopt.org.starbucks.ui.mymenu
 
+import sopt.org.starbucks.data.model.MenuDetailModel
 import sopt.org.starbucks.ui.mymenu.component.DrinkSize
 import sopt.org.starbucks.ui.mymenu.component.TabType
 
 data class EditMenuUiState(
-    val menu: MenuDetail = MenuDetail(),
+    val menu: MenuDetailModel = MenuDetailModel(),
     val selectedTab: TabType = TabType.ICED,
     val selectedSize: DrinkSize = DrinkSize.TALL,
     val isPersonalCupChecked: Boolean = false
